+++ CHANGELOG.md
@@ -17,8 +17,8 @@ Changelog for NeoFS Node
 - `session create-v2` cli command to create new session token v2 (#3750)
 - SN now support raw GET/HEAD/RANGE requests in EC containers (#3756)
 - IR now serves `setAttribute` and `removeAttribute` methods of Container contract (#3733)
-- SN now serves `ContainerService`'s `SetAttribute` and `RemoveAttribute` RPC (#3733)
-- CLI `set-attribute` and `remove-attribute commands to `container` section (#3733)
+- SN now serves `ContainerService`'s `SetAttribute` and `RemoveAttribute` RPC (#3733, #3784)
+- CLI `set-attribute` and `remove-attribute commands to `container` section (#3733, #3784)
 - Size-split object is now available even when LINK is unavailable in EC container (#3754)
 - `neofs-adm balance container-status` command (#3693)
 
@@ -56,7 +56,7 @@ Changelog for NeoFS Node
 
 ### Updated
 - `github.com/nspcc-dev/neofs-contract` module to `v0.26.0` (#3670, #3746, #3733, #3780, #3782)
-- `github.com/nspcc-dev/neofs-sdk-go` module to `v1.0.0-rc.16.0.20260119140653-e47bb5276235` (#3711, #3750, #3733, #3775, #3772, #3787)
+- `github.com/nspcc-dev/neofs-sdk-go` module to `v1.0.0-rc.16.0.20260126114348-87674e46ef14` (#3711, #3750, #3733, #3775, #3772, #3787, #3784)
 - `github.com/nspcc-dev/locode-db` module to `v0.8.2` (#3729)
 - `github.com/nspcc-dev/neo-go` module to `v0.116.0` (#3733, #3769, #3779)